+++ src/TVTest/TVTest_Image/JPEG.cpp
@@ -76,7 +76,7 @@ static boolean JPEGEmptyOutputBuffer(j_compress_ptr cinfo)
 static void JPEGTermDestination(j_compress_ptr cinfo)
 {
 	JPEGDestinationInfo *pInfo=(JPEGDestinationInfo*)cinfo->dest;
-	DWORD Size=JPEG_OUTPUT_BUFFER_SIZE-pInfo->dest.free_in_buffer;
+	DWORD Size=JPEG_OUTPUT_BUFFER_SIZE-(DWORD)pInfo->dest.free_in_buffer;
 
 	if (Size>0) {
 		DWORD Write;
